Extract analyses endpoints (#3546)
