+++ cg/server/app.py
@@ -8,6 +8,7 @@ from sqlalchemy.orm import scoped_session
 
 from cg.server import admin, api, ext, invoices
 from cg.server.app_config import app_config
+from cg.server.endpoints.analyses import ANALYSES_BLUEPRINT
 from cg.server.endpoints.flow_cells import FLOW_CELLS_BLUEPRINT
 from cg.server.endpoints.orders import ORDERS_BLUEPRINT
 from cg.server.endpoints.applications import APPLICATIONS_BLUEPRINT
@@ -98,6 +99,8 @@ def _register_blueprints(app: Flask):
     app.register_blueprint(SAMPLES_BLUEPRINT)
     app.register_blueprint(POOLS_BLUEPRINT)
     app.register_blueprint(FLOW_CELLS_BLUEPRINT)
+    app.register_blueprint(ANALYSES_BLUEPRINT)
+
     _register_admin_views()
 
     ext.csrf.exempt(api.BLUEPRINT)
@@ -107,6 +110,7 @@ def _register_blueprints(app: Flask):
     ext.csrf.exempt(ORDERS_BLUEPRINT)
     ext.csrf.exempt(POOLS_BLUEPRINT)
     ext.csrf.exempt(FLOW_CELLS_BLUEPRINT)
+    ext.csrf.exempt(ANALYSES_BLUEPRINT)
 
     @app.route("/")
     def index():
